+++ modules/video/src/trackerMIL.cpp
@@ -42,164 +42,10 @@
 //M*/
 
 #include "precomp.hpp"
+#include "trackerMILModel.hpp"
 
 namespace cv
 {
-
-/**
- * Implementation of the target state for MIL algorithm
- */
-class TrackerMILTargetState : public TrackerTargetState
-{
-
-public:
-	/**
-	 * \brief Constructor
-	 * \param position Top left corner of the bounding box
-	 * \param width Width of the bounding box
-	 * \param height Height of the bounding box
-	 * \param foreground label for target or background
-	 * \param HAARFeatures features extracted
-	 */
-	TrackerMILTargetState( const Point2f& position, int targetWidth, int targetHeight, bool foreground, const Mat& HAARFeatures )
-	{
-		setTargetPosition( position );
-		setWidth( targetWidth );
-		setHeight( targetHeight );
-		setTargetFg( foreground );
-		setFeatures( HAARFeatures );
-	};
-
-	/**
-	 * \brief Destructor
-	 */
-	~TrackerMILTargetState(){};
-
-	/**
-	 * setters and getters
-	 */
-	inline void setWidth( int targetWidth ){ width = targetWidth; };
-	inline void setHeight( int targetHeight ){ height = targetHeight; };
-	inline void setTargetFg( bool foreground ){ isTarget = foreground; };
-	inline void setFeatures( const Mat& HAARFeatures ){ features = HAARFeatures; };
-	inline int getWidth() const { return width; };
-	inline int getHeight() const { return height; };
-	inline bool isTargetFg() const { return isTarget; };
-	inline Mat getFeatures() const { return features; };
-
-private:
-	int width;
-	int height;
-	bool isTarget;
-	Mat features;
-};
-
-
-/**
- * \brief Implementation of TrackerModel for MIL algorithm
- */
-class TrackerMILModel : public TrackerModel
-{
-public:
-	enum
-	{
-		MODE_POSITIVE    = 1,  // mode for positive features
-		MODE_NEGATIVE    = 2   // mode for negative features
-	};
-
-	/**
-	 * \brief Constructor
-	 */
-	TrackerMILModel();
-
-	/**
-	 * \brief Destructor
-	 */
-	~TrackerMILModel(){};
-
-	/**
-	 * \brief Set the mode
-	 */
-	void setMode( int trainingMode, const std::vector<Mat>& samples );
-
-protected:
-	void modelEstimationImpl( const std::vector<Mat>& responses );
-	void modelUpdateImpl();
-
-private:
-	int mode;
-	std::vector<Mat> currentSample;
-};
-
-/**
- * TrackerMILModel
- */
-TrackerMILModel::TrackerMILModel()
-{
-	currentSample.clear();
-	mode = MODE_POSITIVE;
-}
-
-void TrackerMILModel:: modelEstimationImpl( const std::vector<Mat>& responses )
-{
-	if( currentSample.size() == 0)
-	{
-		CV_Error(-1, "The samples in Model estimation are empty");
-		return;
-	}
-
-
-	for( size_t i = 0; i <  responses.size(); i++ )
-	{
-		//for each column (one sample) there are #num_feature
-		//get informations from currentSample
-		for( int j = 0; j < responses.at(i).cols; j++)
-		{
-
-			Size currentSize;
-			Point currentOfs;
-			currentSample.at(j).locateROI( currentSize, currentOfs );
-			bool foreground;
-			if( mode == MODE_POSITIVE )
-			{
-				foreground = true;
-			}
-			else
-			{
-				foreground = false;
-			}
-
-			//get the column of the HAAR responses
-			Mat singleResponse = responses.at(i).col(j);
-
-			//create the state
-			Ptr<TrackerMILTargetState> currentState = new TrackerMILTargetState( currentOfs,
-					currentSize.width,
-					currentSize.height,
-					foreground,
-					singleResponse);
-
-			currentConfidenceMap.push_back( std::make_pair(currentState, 0) );
-
-		}
-
-
-	}
-
-}
-
-void TrackerMILModel::modelUpdateImpl()
-{
-
-}
-
-void TrackerMILModel::setMode( int trainingMode, const std::vector<Mat>& samples )
-{
-	currentSample.clear();
-	currentSample = samples;
-
-	mode = trainingMode;
-}
 
 RNG TrackerMIL::rng;
 /*
