+++ modules/video/src/trackerMILModel.hpp
@@ -1,0 +1,143 @@
+/*M///////////////////////////////////////////////////////////////////////////////////////
+//
+//  IMPORTANT: READ BEFORE DOWNLOADING, COPYING, INSTALLING OR USING.
+//
+//  By downloading, copying, installing or using the software you agree to this license.
+//  If you do not agree to this license, do not download, install,
+//  copy or use the software.
+//
+//
+//                          License Agreement
+//                For Open Source Computer Vision Library
+//
+// Copyright (C) 2000-2008, Intel Corporation, all rights reserved.
+// Copyright (C) 2009, Willow Garage Inc., all rights reserved.
+// Copyright (C) 2013, OpenCV Foundation, all rights reserved.
+// Third party copyrights are property of their respective owners.
+//
+// Redistribution and use in source and binary forms, with or without modification,
+// are permitted provided that the following conditions are met:
+//
+//   * Redistribution's of source code must retain the above copyright notice,
+//     this list of conditions and the following disclaimer.
+//
+//   * Redistribution's in binary form must reproduce the above copyright notice,
+//     this list of conditions and the following disclaimer in the documentation
+//     and/or other materials provided with the distribution.
+//
+//   * The name of the copyright holders may not be used to endorse or promote products
+//     derived from this software without specific prior written permission.
+//
+// This software is provided by the copyright holders and contributors "as is" and
+// any express or implied warranties, including, but not limited to, the implied
+// warranties of merchantability and fitness for a particular purpose are disclaimed.
+// In no event shall the Intel Corporation or contributors be liable for any direct,
+// indirect, incidental, special, exemplary, or consequential damages
+// (including, but not limited to, procurement of substitute goods or services;
+// loss of use, data, or profits; or business interruption) however caused
+// and on any theory of liability, whether in contract, strict liability,
+// or tort (including negligence or otherwise) arising in any way out of
+// the use of this software, even if advised of the possibility of such damage.
+//
+//M*/
+
+#ifndef __OPENCV_TRACKER_MIL_MODEL_HPP__
+#define __OPENCV_TRACKER_MIL_MODEL_HPP__
+
+#include "opencv2/core.hpp"
+
+namespace cv
+{
+
+/**
+ * Implementation of the target state for MIL algorithm
+ */
+class TrackerMILTargetState : public TrackerTargetState
+{
+
+public:
+	/**
+	 * \brief Constructor
+	 * \param position Top left corner of the bounding box
+	 * \param width Width of the bounding box
+	 * \param height Height of the bounding box
+	 * \param foreground label for target or background
+	 * \param HAARFeatures features extracted
+	 */
+	TrackerMILTargetState( const Point2f& position, int targetWidth, int targetHeight, bool foreground, const Mat& HAARFeatures )
+	{
+		setTargetPosition( position );
+		setWidth( targetWidth );
+		setHeight( targetHeight );
+		setTargetFg( foreground );
+		setFeatures( HAARFeatures );
+	};
+
+	/**
+	 * \brief Destructor
+	 */
+	~TrackerMILTargetState(){};
+
+	/**
+	 * setters and getters
+	 */
+	inline void setWidth( int targetWidth ){ width = targetWidth; };
+	inline void setHeight( int targetHeight ){ height = targetHeight; };
+	inline void setTargetFg( bool foreground ){ isTarget = foreground; };
+	inline void setFeatures( const Mat& HAARFeatures ){ features = HAARFeatures; };
+	inline int getWidth() const { return width; };
+	inline int getHeight() const { return height; };
+	inline bool isTargetFg() const { return isTarget; };
+	inline Mat getFeatures() const { return features; };
+
+
+private:
+	int width;
+	int height;
+	bool isTarget;
+	Mat features;
+};
+
+
+/**
+ * \brief Implementation of TrackerModel for MIL algorithm
+ */
+class TrackerMILModel : public TrackerModel
+{
+public:
+	enum
+	{
+		MODE_POSITIVE    = 1,  // mode for positive features
+		MODE_NEGATIVE    = 2   // mode for negative features
+	};
+
+	/**
+	 * \brief Constructor
+	 */
+	TrackerMILModel();
+
+	/**
+	 * \brief Destructor
+	 */
+	~TrackerMILModel(){};
+
+	/**
+	 * \brief Set the mode
+	 */
+	void setMode( int trainingMode, const std::vector<Mat>& samples );
+
+protected:
+	void modelEstimationImpl( const std::vector<Mat>& responses );
+	void modelUpdateImpl();
+
+private:
+	int mode;
+	std::vector<Mat> currentSample;
+};
+
+
+
+
+} /* namespace cv */
+
+#endif
